refactor: 설정 화면 UI 컴포넌트 분리 및 구조 개선
`SettingsScreen` 내부에 정의되어 있던 각 설정 섹션 UI를 별도의 Composable 함수로 분리하여 재사용성과 가독성을 높였습니다. 기존의 모든 기능은 유지되며, UI 로직이 각 카드별 파일로 이동되었습니다.

- **`SettingsScreen.kt`**:
    - 기존에 내장되어 있던 카드 UI들을 새로 생성된 Composable 함수로 대체하여 코드를 간소화했습니다. (`MeasurementGuidanceCard`, `BaselineSettingsCard`, `RelativeDisplayCard`, `CalibrationFineTuneCard`, `AutoTimerCard`)
    - `SettingsTopBar`를 호출하여 상단 앱 바를 표시하도록 변경했습니다.

- **`component/settings` 패키지 신설**:
    - **`AutoTimerCard.kt`**: 자동 타이머 설정 UI를 포함합니다.
    - **`BaselineSettingsCard.kt`**: 기준점 설정 및 자동 측정 UI를 포함합니다.
    - **`CalibrationFineTuneCard.kt`**: 절대 보정 미세 조정 슬라이더 및 버튼 UI를 포함합니다.
    - **`MeasurementGuidanceCard.kt`**: 측정 안내 및 현재 소음 표시 UI를 포함합니다.
    - **`RelativeDisplayCard.kt`**: 상대 표시(ΔdB) 모드 스위치 UI를 포함합니다.
    - **`SettingsTopBar.kt`**: "환경 보정 센터" 제목과 실시간 데시벨을 표시하는 상단 바 UI를 포함합니다.
